refactor: flatten component structure, remove atomic design folders
