+++ src/components/Example/index.ts
